refactor(shared): split flat SSE event map into domain-owned modules (resolves #327)
