+++ apps/web/src/stores/stream-store.ts
@@ -2,7 +2,7 @@ import { create } from 'zustand';
 
 import type { StreamErrorDetails } from '@stitch/shared/chat/errors';
 import type { LanguageModelUsage, LanguageModelV3Source } from '@stitch/shared/chat/messages';
-import type { PartDelta, PartUpdate, ToolCallStatus } from '@stitch/shared/chat/realtime';
+import type { PartDelta, PartUpdate, ToolCallStatus } from '@stitch/shared/chat/stream-events';
 
 import { serverFetch } from '@/lib/api';
