+++ packages/server/src/question/service.test.ts
@@ -1,8 +1,8 @@
 import { afterEach, beforeEach, describe, expect, test } from 'bun:test';
 import { eq } from 'drizzle-orm';
 
-import type { SseEventName, SseEventPayloadMap } from '@stitch/shared/chat/realtime';
 import type { PrefixedString } from '@stitch/shared/id';
+import type { SseEventName, SseEventPayloadMap } from '@stitch/shared/realtime';
 
 import { getDb } from '@/db/client.js';
 import { questions } from '@/db/schema/questions.js';
