+++ packages/shared/package.json
@@ -8,7 +8,8 @@
     "./automations/types": "./src/automations/types.ts",
     "./chat/errors": "./src/chat/errors.ts",
     "./chat/messages": "./src/chat/messages.ts",
-    "./chat/realtime": "./src/chat/realtime.ts",
+    "./chat/stream-events": "./src/chat/stream-events.ts",
+    "./chat/session-events": "./src/chat/session-events.ts",
     "./id": "./src/id/index.ts",
     "./ipc/types": "./src/ipc/types.ts",
     "./liquid-ui/catalog": "./src/liquid-ui/catalog.ts",
@@ -17,11 +18,16 @@
     "./liquid-ui/schema": "./src/liquid-ui/schema.ts",
     "./mcp/types": "./src/mcp/types.ts",
     "./permissions/types": "./src/permissions/types.ts",
+    "./permissions/events": "./src/permissions/events.ts",
     "./providers/catalog": "./src/providers/catalog.ts",
     "./providers/model-visibility": "./src/providers/model-visibility.ts",
     "./providers/types": "./src/providers/types.ts",
     "./questions/types": "./src/questions/types.ts",
+    "./questions/events": "./src/questions/events.ts",
     "./recordings/types": "./src/recordings/types.ts",
+    "./recordings/events": "./src/recordings/events.ts",
+    "./recordings/meeting-ipc": "./src/recordings/meeting-ipc.ts",
+    "./realtime": "./src/realtime.ts",
     "./settings/types": "./src/settings/types.ts",
     "./shortcuts/types": "./src/shortcuts/types.ts",
     "./todos/types": "./src/todos/types.ts",
